requirements.txt
# 多模态情感分析系统 - 依赖清单

# =============================================================================
# 深度学习框架
# =============================================================================
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# =============================================================================
# 预训练模型与NLP
# =============================================================================
transformers>=4.30.0
tokenizers>=0.13.0
sentencepiece>=0.1.99

# =============================================================================
# 音频处理
# =============================================================================
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# =============================================================================
# 图像与视频处理
# =============================================================================
opencv-python>=4.8.0
pillow>=10.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# =============================================================================
# 数据处理
# =============================================================================
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# =============================================================================
# 机器学习工具
# =============================================================================
scikit-learn>=1.3.0

# =============================================================================
# 可视化
# =============================================================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# =============================================================================
# Web界面
# =============================================================================
streamlit>=1.28.0
# 或者使用 Gradio (二选一)
# gradio>=4.0.0

# =============================================================================
# 工具库
# =============================================================================
tqdm>=4.65.0
tensorboard>=2.13.0
pyyaml>=6.0

# 可选：实验追踪
# wandb>=0.15.0

# =============================================================================
# 数据集（可选）
# =============================================================================
# mmsdk>=0.1.0  # CMU-MOSEI数据集下载
# kaggle>=1.15.0  # Kaggle数据集下载

# =============================================================================
# 测试
# =============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0

# =============================================================================
# 代码质量（可选）
# =============================================================================
# black>=23.7.0
# flake8>=6.1.0
# isort>=5.12.0
